+++ src/operations/Completion.cpp
@@ -241,6 +241,26 @@ static bool deprecated(const Luau::AutocompleteEntry& entry, std::optional<lsp::
     if (entry.deprecated)
         return true;
 
+    // TODO: unnecessary once https://github.com/luau-lang/luau/issues/2158 is fixed
+    if (entry.type)
+    {
+        const auto ty = Luau::follow(*entry.type);
+        if (const auto ftv = Luau::get<Luau::FunctionType>(ty); ftv && ftv->isDeprecatedFunction)
+            return true;
+        else if (const auto itv = Luau::get<Luau::IntersectionType>(ty))
+        {
+            const auto allDeprecated = std::all_of(itv->parts.begin(), itv->parts.end(),
+                [](const auto& part)
+                {
+                    const auto partTy = Luau::follow(part);
+                    const auto ftv = Luau::get<Luau::FunctionType>(partTy);
+                    return ftv && ftv->isDeprecatedFunction;
+                });
+            if (allDeprecated)
+                return true;
+        }
+    }
+
     if (documentation)
         if (documentation->value.find("@deprecated") != std::string::npos)
             return true;