+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed non-adjacent comments (such as section headers separated by blank lines) being incorrectly included in hover documentation ([#310](https://github.com/JohnnyMorganz/luau-lsp/issues/310))
+- Function entries correctly show as deprecated in autocomplete if they are marked with `@deprecated` attribute ([#1302](https://github.com/JohnnyMorganz/luau-lsp/issues/1302))
 
 ## [1.58.0] - 2025-12-14